Quick Access FileDialog;  (BUGGY) Fractional rate envelopes; Improved bitwiseOp fidelity * FileDialogs now have a special menu to access MRUDs and Favorites * HQ Operators can specify a rate in decimal -!globals.gd:  reparent node -PhaseEngine.csproj:  Allow unsafe code to perform unchecked direct casting -Tools.cs:  (UNSAFE) Method to convert boolean to byte. Used by Modulo funcs in BitwiseOp to eliminate divide-by-zero without branching -Channel.cs:  Force filter func to be in "sane" range on changed opIntent, as target method expects -Envelope.cs:  Extended `aux_func` to 32-bits to support HQ rate scaling -Operator.cs:  Bitwise Modulo funcs now support full 16-bit range. New div-by-zero check only changes right operand to 1 if its value is 0. -OperatorHQ.cs:  Rate envelopes use `aux_func` to extend fixed-point fractional precision to 8 bits.  (MAY CHANGE IN FUTURE) -Voice.cs:  Helper method to set an EG rate for an HQ Operator in its native fixed-precision extension format;  `aux_func` zeroed when intent changed to `FM_HQ` to prevent erroneous rate values -io/SYX.cs:  Support HQ Operator rate value extensions. Implement KSR.  Begin process to implement KSL. -Test2.cs:  Bus to Voice.cs helper method for setting extended rate envelopes -ui/panels/BasePanel.gd:  Added mod wheel to potential future bind types -ui/panels/BitwiseOpPanel.tscn:  Increase mult range to 32 from 16 -ui/panels/EGPanel.gd:  Alter rate sliders to support fractional values if our operator is HQ -ui/panels/EGPanel.gd:  Increase mult range to 32 from 16 -ui/responseCurve/btnPresets.gd:  Change reset to default key combination to prevent accidental presses while playing notes on PC keyboard -ui/voice_io/VoiceIO.tscn:  Default to showing hidden files on FileDialogs. Add Quick Access menu. -ui/voice_io/VoiceIODialogs.gd:  hook the Quick Access menu to the file dialogs by adding a new button. +ui/main/QuickAccess.gd:  Quick Access menu code +ui/main/QuickAccess.tscn: Quick Access menu popup +ui/main/QuickAccessButton.tscn:  Button for creating Quick Access popup inside a FileDialog

+++ PhaseEngine/FMCore/Channel.cs
@@ -284,7 +284,7 @@ namespace PhaseEngine
                             var f = new Filter();
                             ops[i] = f;
                             f.eg.Configure(voice.egs[i]);
-                            f.SetOscillatorType(f.eg.aux_func);  //TODO:  Check if values out of range cause the filter to freak out
+                            f.SetOscillatorType((byte)f.eg.aux_func);  //TODO:  Check if values out of range cause the filter to freak out
                             f.wavetable = voice.wavetable;
                             break;
                         case OpBase.Intents.BITWISE:  //Extends FM_OP
